Rename reducer -> error-overlay-reducer
