+++ packages/next/client/components/react-dev-overlay/hot-reloader.tsx
@@ -20,7 +20,7 @@ import type { DispatchFn } from './client'
 import stripAnsi from 'next/dist/compiled/strip-ansi'
 import formatWebpackMessages from '../../dev/error-overlay/format-webpack-messages'
 import { useRouter } from '../hooks-client'
-import { errorOverlayReducer } from './internal/reducer'
+import { errorOverlayReducer } from './internal/error-overlay-reducer'
 
 function getSocketProtocol(assetPrefix: string): string {
   let protocol = window.location.protocol
