+++ packages/next/client/components/react-dev-overlay/internal/error-overlay-reducer.ts
